Fixing the way we connect to Mongo through the container
Also created a Singleton object of the db so that it can be updated and changed globally.

A method has been created in the app entry __init__ to connect to Mongo and check the connection otherwise it will log errors.

The changes in the base test case should also properly create the testing database and then drop it once its done.