Adding alloys database endpoint and impl design using abstraction models
It starts with an Alloy schema using the marshmallow package. We then create a concrete MongoAlloys object, which is the data access layer, that provides all the methods from the PyMongo package. We then use an Abstract class which provides the ability to store an adapter, using the adapter design pattern. Finally, we create the Service layer that contains all the application/business logic.